+++ src/runtime.rs
@@ -21,6 +21,7 @@ use crate::{Descriptor, Error, Miniscript, Policy, Result, Scope};
 pub enum Value {
     PubKey(DescriptorPublicKey),
     Hash(Vec<u8>),
+    Bytes(Vec<u8>),
     Number(usize),
     DateTime(String),
     Duration(ast::Duration),
@@ -48,6 +49,7 @@ impl_from_variant!(Script, Value);
 impl_from_variant!(Address, Value);
 impl_from_variant!(Function, Value);
 impl_from_variant!(Array, Value);
+impl_from_variant!(Vec<u8>, Value, Bytes);
 impl_from_variant!(Network, Value);
 impl_from_variant!(usize, Value, Number);
 
@@ -251,6 +253,7 @@ impl Evaluate for Expr {
             // Atoms
             Expr::PubKey(x) => Value::PubKey(x.parse()?),
             Expr::Hash(x) => Value::Hash(Vec::from_hex(&x)?),
+            Expr::Bytes(x) => Value::Bytes(x.clone()),
             Expr::Number(x) => Value::Number(*x),
             Expr::Duration(x) => Value::Duration(x.clone()),
             Expr::DateTime(x) => Value::DateTime(x.clone()),
@@ -359,14 +362,12 @@ impl TryFrom<Value> for Script {
                 Ok(v.into_desc()?.script_code(ctx))
             }
             Value::Number(n) => Ok(ScriptBuilder::new().push_int(n as i64).into_script()),
+            Value::Bytes(bytes) => Ok(ScriptBuilder::new().push_slice(&bytes).into_script()),
             Value::PubKey(desc_pubkey) => {
                 let pubkey = desc_pubkey.to_public_key(get_descriptor_ctx(0));
                 Ok(ScriptBuilder::new().push_key(&pubkey).into_script())
             }
-            Value::Hash(hash) => {
-                let script = ScriptBuilder::new().push_slice(&hash).into_script();
-                Ok(script.into())
-            }
+            Value::Hash(hash) => Ok(ScriptBuilder::new().push_slice(&hash).into_script()),
             Value::Duration(dur) => {
                 let seq_num = time::duration_to_seq(&dur)?;
                 Ok(ScriptBuilder::new().push_int(seq_num as i64).into_script())
@@ -399,6 +400,7 @@ macro_rules! impl_hash_conv {
             fn try_from(value: Value) -> Result<Self> {
                 match value {
                     Value::Hash(h) => Ok(Self::from_slice(&h)?),
+                    Value::Bytes(b) => Ok(Self::from_slice(&b)?),
                     v => Err(Error::NotHash(v)),
                 }
             }
@@ -450,6 +452,7 @@ impl fmt::Display for Value {
             Value::DateTime(x) => write!(f, "{}", x),
             Value::Duration(x) => write!(f, "{:?}", x),
             Value::Hash(x) => write!(f, "{}", x.to_hex()),
+            Value::Bytes(x) => write!(f, "{}", x.to_hex()),
             Value::Policy(x) => write!(f, "{}", x),
             Value::WithProb(p, x) => write!(f, "{}@{}", p, x),
             Value::Miniscript(x) => write!(f, "{}", x),